requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.12.0
faiss-cpu>=1.8.0
tiktoken>=0.6.0
PyPDF2>=3.0.0
python-docx>=1.1.0
numpy>=1.26.0
python-dotenv>=1.0.0